requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-ppc64le
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
brotli=1.1.0=ha17a0cc_1
brotli-bin=1.1.0=ha17a0cc_1
brotli-python=1.1.0=py310h9de49d8_1
bzip2=1.0.8=ha17a0cc_5
ca-certificates=2023.7.22=h0f6029e_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.16.0=py310he93fcb1_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
cuda=11.8.0=0
cuda-cccl=11.8.89=0
cuda-command-line-tools=11.8.0=0
cuda-compiler=11.8.0=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cuobjdump=11.8.86=0
cuda-cupti=11.8.87=0
cuda-cuxxfilt=11.8.86=0
cuda-documentation=12.3.52=0
cuda-driver-dev=12.3.52=0
cuda-gdb=12.3.52=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.3.52=0
cuda-nsight-compute=12.3.0=0
cuda-nvcc=11.8.89=0
cuda-nvdisasm=12.3.52=0
cuda-nvml-dev=12.3.52=0
cuda-nvprof=12.3.52=0
cuda-nvprune=11.8.86=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-nvvp=12.3.52=0
cuda-profiler-api=12.3.52=0
cuda-runtime=11.8.0=0
cuda-sanitizer-api=12.3.52=0
cuda-toolkit=11.8.0=0
cuda-tools=11.8.0=0
cuda-visual-tools=11.8.0=0
cycler=0.12.1=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
fonttools=4.44.0=py310h05217ba_0
freetype=2.12.1=h5baba08_2
future=0.18.3=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.3.2=pyhd8ed1ab_0
kiwisolver=1.4.5=py310h3cb7a77_1
lcms2=2.15=ha74cc02_3
ld_impl_linux-ppc64le=2.40=hfa29eb9_0
lerc=4.0.0=hbbae597_0
libblas=3.9.0=19_linuxppc64le_openblas
libbrotlicommon=1.1.0=ha17a0cc_1
libbrotlidec=1.1.0=ha17a0cc_1
libbrotlienc=1.1.0=ha17a0cc_1
libcblas=3.9.0=19_linuxppc64le_openblas
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.4.52=0
libcurand-dev=10.3.4.52=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.19=ha17a0cc_0
libexpat=2.5.0=h883269e_1
libffi=3.4.2=h4e0d66e_5
libgcc-ng=13.2.0=h31ea8bd_2
libgfortran-ng=13.2.0=hfdc3801_2
libgfortran5=13.2.0=h7bcb406_2
libgomp=13.2.0=h31ea8bd_2
libjpeg-turbo=3.0.0=ha17a0cc_1
liblapack=3.9.0=19_linuxppc64le_openblas
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnsl=2.0.1=ha17a0cc_0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libopenblas=0.3.24=pthreads_h3f0fafc_0
libpng=1.6.39=hcc10993_0
libprotobuf=3.20.3=h26001cd_0
libsqlite=3.44.0=hd4bbf49_0
libstdcxx-ng=13.2.0=h0849fd4_2
libtiff=4.6.0=hb1021a9_2
libuuid=2.38.1=h4194056_0
libwebp-base=1.3.2=ha17a0cc_0
libxcb=1.15=h4194056_0
libzlib=1.2.13=ha17a0cc_5
markupsafe=2.1.3=py310h6ed3b71_1
matplotlib-base=3.5.3=py310h34844f2_2
mne-base=1.5.1=pyh57928b3_0
mne-icalabel=0.6.0.dev0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=h8800142_2
ninja=1.11.1=h06f31f1_0
nsight-compute=2023.3.0.12=0
numpy=1.26.0=py310h877c07d_0
openjpeg=2.5.0=h1c3486c_3
openssl=3.1.4=ha17a0cc_0
packaging=23.2=pyhd8ed1ab_0
pillow=10.1.0=py310hf64def1_0
pip=23.3.1=pyhd8ed1ab_0
platformdirs=3.11.0=pyhd8ed1ab_0
pooch=1.8.0=pyhd8ed1ab_0
psutil=5.9.6=pypi_0
pthread-stubs=0.4=h339bb43_1001
pycparser=2.21=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.13=h4005451_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytorch=1.13.1=cpu_py310hc26b713_0
pytorch-cuda=11.8=h8dd9ede_2
pyyaml=6.0.1=py310h05217ba_1
readline=8.2=h0b9b154_1
requests=2.31.0=pyhd8ed1ab_0
scikit-learn=1.3.2=py310h42bb3b2_1
scipy=1.11.3=py310h877c07d_1
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.13=hd4bbf49_0
tqdm=4.66.1=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
unicodedata2=15.1.0=py310h05217ba_0
urllib3=2.0.7=pyhd8ed1ab_0
wheel=0.41.3=pyhd8ed1ab_0
xorg-libxau=1.0.11=ha17a0cc_0
xorg-libxdmcp=1.1.3=h4e0d66e_0
xz=5.2.6=hb283c62_0
yaml=0.2.5=h4e0d66e_2
zstd=1.5.5=h7292585_0
